Fixed detection of DRM in audio books
Turns out, all audio books at Kobo have DRM at the moment in the sense
that they cannot be downloaded or listened to with anything but the Kobo
iOS or Android apps. This script was modified to detect audio books as
such now.

Oher changes were made. The checks are now ALL done for each book
because we may have multiple, conflicting indications in one book. We
now report the "worst" thing observed about a book with unknown being
the default if we can't tell anything.

Because applying many regexps like this is even slower, the graying out
of the book names was pulled out into a separate, earlier loop.